Removed call to obsolete/removed constructor in Newtonsoft.Json
The usage of the shareCache argument has been obsolete for a long time in
Newtonsoft.Json and was finally removed in the latest versions. Newtonsoft
recommends that caching is done within the application outside of this
library. If this automatic caching is important stick to the older
versions as it will no longer have any effect.